+++ src/Taxes/Exceptions/InvalidTaxConfigurationException.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Contains the InvalidTaxConfigurationException class.
+ *
+ * @copyright   Copyright (c) 2023 Vanilo UG
+ * @author      Attila Fulop
+ * @license     MIT
+ * @since       2023-03-26
+ *
+ */
+
+namespace Vanilo\Taxes\Exceptions;
+
+use RuntimeException;
+
+class InvalidTaxConfigurationException extends RuntimeException
+{
+}