+++ src/Adjustments/Changelog.md
@@ -2,6 +2,11 @@
 
 ## 3.x Series
 
+## Unreleased
+##### 2023-XX-YY
+
+- Added the `SimpleTax` adjuster (it calculates the tax amount based on a flat rate in %)
+
 ## 3.6.0
 ##### 2023-03-07